+++ src/lib/st-textarea/st-textarea.module.ts
@@ -12,11 +12,11 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
-import { StTooltipModule } from '../st-tooltip/st-tooltip.module';
 import { StTextareaComponent } from './st-textarea.component';
+import { StLabelModule } from '../st-label/st-label.module';
 
 @NgModule({
-   imports: [CommonModule, FormsModule, ReactiveFormsModule, StTooltipModule],
+   imports: [CommonModule, FormsModule, ReactiveFormsModule, StLabelModule],
    declarations: [StTextareaComponent],
    exports: [StTextareaComponent]
 })